Factor out BackendPersister
